Make github-clone-add-existing-remote use source
Typically, users of github-clone-add-existing-remote will want to clone a
fork of the original repository, even if the remote they select is a
fork. This branch makes it so that the source repository is used to
collect forks when calling in to the github api using gh.el. This
behavior can be controlled with the new second argument to the function
use-source which will disable this behavior.